+++ src/Temporalio/Worker/WorkflowInstance.cs
@@ -32,7 +32,7 @@ namespace Temporalio.Worker
     /// <summary>
     /// Instance of a workflow execution.
     /// </summary>
-    internal class WorkflowInstance : TaskScheduler, IWorkflowInstance, IWorkflowContext, IWorkflowCodecHelperInstance
+    internal partial class WorkflowInstance : TaskScheduler, IWorkflowInstance, IWorkflowContext, IWorkflowCodecHelperInstance
     {
         private static readonly string[] Newlines = new[] { "\r", "\n", "\r\n" };
         private static readonly AsyncLocal<WorkflowUpdateInfo> CurrentUpdateInfoLocal = new();
@@ -2102,7 +2102,7 @@ namespace Temporalio.Worker
         /// <summary>
         /// Workflow outbound implementation.
         /// </summary>
-        internal class OutboundImpl : WorkflowOutboundInterceptor
+        internal partial class OutboundImpl : WorkflowOutboundInterceptor
         {
             private readonly WorkflowInstance instance;
 
@@ -2682,15 +2682,34 @@ namespace Temporalio.Worker
                         new CanceledFailureException("Nexus operation cancelled before scheduled"));
                 }
 
-                // TODO: Scope the generated System Nexus support converter context around this
-                // operation converter once the generated support file is ingested into the SDK.
+                ISerializationContext? serializationContext = null;
+                if (SystemNexusPayloadVisitor.IsSystemEndpoint(input.ClientOptions.Endpoint) &&
+                    input.Arg is { } arg &&
+                    NexgenOperationRegistry.Operations.TryGetValue(
+                        (input.Service, input.OperationName), out var operationInfo))
+                {
+                    serializationContext = operationInfo.SerializationContext?.Invoke(arg);
+                }
+
+                var payloadConverter = instance.payloadConverterNoContext;
+                var failureConverter = instance.failureConverterNoContext;
+                if (serializationContext != null)
+                {
+                    if (payloadConverter is IWithSerializationContext<IPayloadConverter> payloadWithContext)
+                    {
+                        payloadConverter = payloadWithContext.WithSerializationContext(serializationContext);
+                    }
+                    if (failureConverter is IWithSerializationContext<IFailureConverter> failureWithContext)
+                    {
+                        failureConverter = failureWithContext.WithSerializationContext(serializationContext);
+                    }
+                }
+
                 var systemNexusPayloadConverter = SystemNexusPayloadVisitor.IsSystemEndpoint(
                     input.ClientOptions.Endpoint) ?
-                    new SystemNexusPayloadConverter(
-                        instance.payloadConverterNoContext,
-                        instance.failureConverterNoContext) : null;
+                    new SystemNexusPayloadConverter(payloadConverter, failureConverter) : null;
                 var operationPayloadConverter =
-                    systemNexusPayloadConverter ?? instance.payloadConverterNoContext;
+                    systemNexusPayloadConverter ?? payloadConverter;
 
                 var seq = ++instance.nexusOperationCounter;
                 var inputPayload = operationPayloadConverter.ToPayload(input.Arg);
@@ -2761,15 +2780,12 @@ namespace Temporalio.Worker
                             // there's nothing more we can do here
                             var handle = new NexusWorkflowOperationHandleImpl<TResult>(
                                 operationPayloadConverter,
-                                // TODO(cretz): Support Nexus serialization context, ideally not
-                                // creating failure converter with context until actually needed
-                                instance.failureConverterNoContext,
+                                failureConverter,
                                 startRes.HasOperationToken ? startRes.OperationToken : null);
                             if (startRes.Failed is { } syncStartFail)
                             {
-                                // TODO(cretz): Support Nexus serialization context
                                 handleSource.SetException(
-                                    instance.failureConverterNoContext.ToException(
+                                    failureConverter.ToException(
                                         syncStartFail, operationPayloadConverter));
                                 return;
                             }
@@ -2796,6 +2812,17 @@ namespace Temporalio.Worker
                 });
                 return handleSource.Task;
             }
+
+            /// <inheritdoc/>
+            public override Task<NexusWorkflowOperationHandle<TResult>> ScheduleSystemNexusOperationAsync<TResult>(
+                ScheduleSystemNexusOperationInput input) =>
+                ScheduleNexusOperationAsync<TResult>(new(
+                    Service: input.Service,
+                    ClientOptions: new(SystemNexusPayloadVisitor.TemporalSystemEndpoint),
+                    OperationName: input.Operation.Name,
+                    Arg: input.Arg,
+                    Options: new(),
+                    Headers: input.Headers));
 
             private Task SignalExternalWorkflowInternalAsync(
                 ISerializationContext.Workflow serializationContext,
@@ -3090,6 +3117,8 @@ namespace Temporalio.Worker
 
             public override Task<NexusWorkflowOperationHandle<TResult>> StartNexusOperationAsync<TResult>(
                 string operationName, object? arg, NexusWorkflowOperationOptions? options = null) =>
+                SystemNexusPayloadVisitor.IsSystemEndpoint(Options.Endpoint) ?
+                instance.StartSystemNexusOperationAsync<TResult>(Service, operationName, arg) :
                 instance.outbound.Value.ScheduleNexusOperationAsync<TResult>(new(
                     Service: Service,
                     ClientOptions: Options,
@@ -3116,6 +3145,8 @@ namespace Temporalio.Worker
 
             public override Task<NexusWorkflowOperationHandle<TResult>> StartNexusOperationAsync<TResult>(
                 string operationName, object? arg, NexusWorkflowOperationOptions? options = null) =>
+                SystemNexusPayloadVisitor.IsSystemEndpoint(Options.Endpoint) ?
+                instance.StartSystemNexusOperationAsync<TResult>(Service, operationName, arg) :
                 instance.outbound.Value.ScheduleNexusOperationAsync<TResult>(new(
                     Service: Service,
                     ClientOptions: Options,